+++ BitwiseOperator/ArqumShaikh.c
