--- nishanthebbar/ne
@@ -1,13 +1,0 @@
-a=int(input("Please enter the first number"))
-b=int(input("Please enter the second number"))
-x=input("Please Enter operand")
-
-if x=='+':
-    print(a+b)
-elif x=='-':
-    print(a-b)
-elif x=='*':
-    print(a*b)
-elif x=='/':
-    print(a/b)
-
